requirements.txt
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.3
asttokens==2.4.1
async-timeout==4.0.3
asyncpg==0.29.0
beautifulsoup4==4.12.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.109.1
filelock==3.13.1
fonttools==4.47.0
fsspec==2024.2.0
googlesearch-python==1.2.3
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.20.3
idna==3.6
ipykernel==6.28.0
ipython==8.20.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.26.3
packaging==23.2
pandas==2.1.4
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.6.0
pydantic_core==2.16.1
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.11.4
seaborn==0.13.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.29
stack-data==0.6.3
starlette==0.35.1
sympy==1.12
threadpoolctl==3.2.0
tld==0.13
tldextract==5.1.1
tokenizers==0.15.1
torch==2.2.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.37.2
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
uvicorn==0.27.0.post1
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
wordcloud==1.9.3
xgboost==2.0.3
